pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "live-subtitle-overlay"
version = "0.1.0"
description = "Windows live subtitle overlay for foreign-language video playback."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "SmoothShadow" }
]
dependencies = [
  "numpy>=1.26",
  "faster-whisper>=1.1.0",
  "PySide6>=6.8",
  "PyAudioWPatch>=0.2.12.8; sys_platform == 'win32'",
  "webrtcvad-wheels>=2.0.14; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[project.scripts]
live-subtitle-overlay = "live_subtitle_overlay.app:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
